Moves components into separate files
