+++ src/components/Badge/index.js
@@ -1,0 +1,3 @@
+import Badge from './Badge'
+
+export { Badge }
