+++ src/components/SidebarFilterTerms/index.js
@@ -1,0 +1,3 @@
+import SidebarFilterTerms from './SidebarFilterTerms'
+
+export { SidebarFilterTerms }
