+++ src/components/TableRows/index.js
@@ -1,0 +1,3 @@
+import TableRows from './TableRows'
+
+export { TableRows }
